// src/app/interfaces/expenses.interface.ts
export interface Expenses {
    id: number;
    date_request: string;
    name: string;
    surname: string;
    department: string;
    project_code: string;
    reason: string;
    date_expense: string;
    type: string;
    amount: number;
    attached: string;
    departure_date: string;
    return_date: string;
    overnight: number;
    trip_origin: string;
    trip_destination: string;
    means_transport: string;
    if_car_km: number;
    product_link: string;
    provider: string;
    units: number;
    requested_training: string;
    training_hours: string;
    training_schedule: string;
    number_people: number;
    hotel_link: string;
    general_type: string;
    is_accepted: string;
    validator_note: string;
    users_id: number;
}